hotfix: prevent delete game with active (#33)

+++ src/FCG.Catalog.Application/UseCases/Games/Delete/DeleteGameUseCase.cs
@@ -1,6 +1,7 @@
 ﻿using FCG.Catalog.Domain.Abstractions;
 using FCG.Catalog.Domain.Exception;
 using FCG.Catalog.Domain.Repositories.Game;
+using FCG.Catalog.Domain.Repositories.Promotion;
 using FCG.Catalog.Messages;
 using MediatR;
 
@@ -10,26 +11,31 @@ namespace FCG.Catalog.Application.UseCases.Games.Delete
     {
         private readonly IReadOnlyGameRepository _gameRepository;
         private readonly IUnitOfWork _unitOfWork;
+        private readonly IReadOnlyPromotionRepository _readOnlyPromotionRepository;
 
         public DeleteGameUseCase(
             IReadOnlyGameRepository gameRepository,
+            IReadOnlyPromotionRepository readOnlyPromotionRepository,
             IUnitOfWork unitOfWork)
         {
             _gameRepository = gameRepository;
+            _readOnlyPromotionRepository = readOnlyPromotionRepository;
             _unitOfWork = unitOfWork;
         }
 
         public async Task<Unit> Handle(DeleteGameInput request, CancellationToken cancellationToken)
         {
             var game = await _gameRepository.GetByIdAsync(request.Id, cancellationToken);
-
+            var promotions = await _readOnlyPromotionRepository.GetByGameIdAsync(request.Id, cancellationToken);
 
             if (game is null)
             {
                 throw new NotFoundException(ResourceMessages.GameNotFound);
             }
-
-
+            if (promotions.Any())
+            {
+                throw new DomainException(ResourceMessages.GameWithPromotion);
+            }
             await _gameRepository.Delete(game, cancellationToken);
 
             await _unitOfWork.CommitTransactionAsync(cancellationToken);
